Move maths utils into the num_t structure, require explicit conversions

+++ src/player/player_update.cpp
@@ -53,8 +53,8 @@
 void p_PlayerMakeFlySparkle(const Location_t& loc, int Frame)
 {
     NewEffect(EFFID_SPARKLE,
-        newLoc(loc.X - 8 + dRand() * (loc.Width + 16) - 4,
-            loc.Y - 8 + dRand() * (loc.Height + 16)),
+        newLoc(loc.X - 8 + dRand() * ((int)loc.Width + 16) - 4,
+            loc.Y - 8 + dRand() * ((int)loc.Height + 16)),
         1, 0, ShadowMode);
     Effect[numEffects].Location.SpeedX = (dRand() / 2) - 0.25_n;
     Effect[numEffects].Location.SpeedY = (dRand() / 2) - 0.25_n;
